# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/area": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取参数域列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数管理"
                ],
                "summary": "获取参数域列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "是否需要分页， 默认需要， 如果不分页 传 true",
                        "name": "isNotPaging",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页码， 如果不分页 传 0",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "一页大小， 如果不分页 传 0",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "检索内容",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Full"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ParamArea"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改参数域信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数管理"
                ],
                "summary": "修改参数域信息",
                "parameters": [
                    {
                        "description": "修改参数域信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReqModifyArea"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加参数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数管理"
                ],
                "summary": "添加参数",
                "parameters": [
                    {
                        "description": "添加参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReqCreateArea"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
        "/api/area/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取参数域列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数管理"
                ],
                "summary": "获取参数域列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "参数域ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
...
        "/api/param": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取参数列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数管理"
                ],
                "summary": "获取参数列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "是否需要分页， 默认需要， 如果不分页 传 true",
                        "name": "isNotPaging",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页码， 如果不分页 传 0",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "一页大小， 如果不分页 传 0",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询内容",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "参数域ID",
                        "name": "area_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Full"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Param"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改参数信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数管理"
                ],
                "summary": "修改参数信息",
                "parameters": [
                    {
                        "description": "修改参数信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReqModifyParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加参数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数管理"
                ],
                "summary": "添加参数",
                "parameters": [
                    {
                        "description": "添加参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReqCreateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
        "/api/param/{areaid}/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除参数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数管理"
                ],
                "summary": "删除参数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "参数域ID",
                        "name": "areaid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "参数ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Param": {
            "type": "object",
            "properties": {
                "area": {
                    "description": "参数域",
                    "type": "string"
                },
                "area_id": {
                    "description": "参数域id",
                    "type": "integer"
                },
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "参数名称",
                    "type": "string"
                },
                "value": {
                    "description": "参数值",
                    "type": "string"
                }
            }
        },
        "model.ParamArea": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "域名称",
                    "type": "string"
                }
            }
        },
        "model.ReqCreateArea": {
            "type": "object",
            "properties": {
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "域名称",
                    "type": "string"
                }
            }
        },
...
        "model.ReqCreateParam": {
            "type": "object",
            "properties": {
                "area_id": {
                    "description": "参数域id",
                    "type": "integer"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "参数名称",
                    "type": "string"
                },
                "value": {
                    "description": "参数值",
                    "type": "string"
                }
            }
        },
...
        "model.ReqModifyArea": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "域名称",
                    "type": "string"
                }
            }
        },
...
        "model.ReqModifyParam": {
            "type": "object",
            "properties": {
                "area": {
                    "description": "参数域",
                    "type": "string"
                },
                "area_id": {
                    "description": "参数域id",
                    "type": "integer"
                },
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "参数名称",
                    "type": "string"
                },
                "value": {
                    "description": "参数值",
                    "type": "string"
                }
            }
        },
...
        "res.Base": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/res.RCode"
                        }
                    ]
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
        "res.Full": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/res.RCode"
                        }
                    ]
                },
                "data": {
                    "description": "返回数据"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
        "res.RCode": {
            "type": "integer",
            "enum": [
                200,
                201,
                400,
                401,
                404,
                405,
                500,
                501,
                502,
                504,
                505
            ],
            "x-enum-varnames": [
                "OK",
                "OK_CREATE",
                "BAD_REQUEST",
                "UNAUTHORIZED",
                "NOT_FOUND",
                "METHOD_NOT_ALLOWED",
                "INTERNAL_SERVER_ERROR",
                "NOT_IMPLEMENTED",
                "BAD_GATEWAY",
                "GATEWAY_TIMEOUT",
                "HTTP_VERSION_NOT_SUPPORTED"
            ]
        }
    }
}